moved animation class to util
